pyproject.toml
[project]
name = "summit-housing"
version = "0.1.0"
description = "Summit County Housing Analysis Portfolio Project"
authors = [{name = "Brian", email = "brian@example.com"}]
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "streamlit>=1.24.0",
    "plotly>=5.15.0",
    "matplotlib>=3.7.0",
    "shap>=0.44.0",
    "statsmodels>=0.14.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.8.0",
    "beautifulsoup4>=4.12.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "torch>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.0.280",
    "black>=23.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.hatch.build.targets.wheel]
packages = ["src/summit_housing"]
